Fix search index generation
There were a few problems with the indexing and generation.

First, the indexer was indexing the entire project contents instead
of just the articles. I modified the loop so that it skips any resources
that don't have a title, that are not published, or that have
`index: false` specified.

It is more common to want to index the majority of your posts. This way,
you only need to manually specify `index: false` for a small amount of
posts that you want to ignore, as opposed to having to manually specify
`index: true` for hundreds of posts.

Second, the JSON parsing and generation were resulting in a
"nesting too deep" error. To fix this, we need to add
`max_nesting: false` to both the `JSON.parse` call in `indexer.rb` and in
the `JSON.generate` call in `search.json`.